# api/openapi.yaml
openapi: 3.0.3
info:
  title: PhantomStrike API
  description: AI-Native Autonomous Security Testing Platform API
  version: 1.0.0
  contact:
    name: PhantomStrike
    url: https://github.com/ersinkoc/phantomstrike
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /missions:
    get:
      summary: List missions
      description: Returns a list of missions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of missions
          content:
            application/json:
              schema:
                type: object
                properties:
                  missions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Mission"

    post:
      summary: Create mission
      description: Create a new security testing mission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionCreate"
      responses:
        "201":
          description: Mission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}:
    get:
      summary: Get mission
      description: Get mission details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Mission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          description: Mission not found

    put:
      summary: Update mission
      description: Update mission details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionUpdate"
      responses:
        "200":
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"

    delete:
      summary: Delete mission
      description: Delete a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Mission deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Mission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [created, planning, recon, scanning, exploitation, post_exploit, completed, cancelled, failed, paused]
        mode:
          type: string
          enum: [autonomous, guided, manual]
        depth:
          type: string
          enum: [quick, standard, deep, exhaustive]
        target:
          type: object
        phases:
          type: array
          items:
            type: string
        current_phase:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MissionCreate:
      type: object
      required:
        - name
        - target
        - mode
        - depth
      properties:
        name:
          type: string
        description:
          type: string
        target:
          type: object
        mode:
          type: string
          enum: [autonomous, guided, manual]
        depth:
          type: string
          enum: [quick, standard, deep, exhaustive]
    MissionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...